Final Week VIII PINN project submission with comprehensive documentation
- Complete LaTeX report (quadrotor_pinn_report.pdf) with:
  * SREEJITA CHATTERJEE authorship
  * Week VIII report designation
  * Fixed header overlaps and table formatting
  * All green backgrounds removed from plots
- Generated 21 high-quality visualizations (5 summary + 16 detailed)
- Updated Python scripts with white/transparent backgrounds
- Comprehensive project structure documentation
- Removed temporary files and outdated model artifacts
- Clean repository ready for final submission

+++ scripts/enhanced_pinn_model.py
@@ -36,11 +36,11 @@ class EnhancedQuadrotorPINN(nn.Module):
         self.network = nn.Sequential(*layers)
         
         # Physical parameters with tighter bounds
-        self.Jxx = nn.Parameter(torch.tensor(6.86e-5, dtype=torch.float32))
-        self.Jyy = nn.Parameter(torch.tensor(9.2e-5, dtype=torch.float32))  
-        self.Jzz = nn.Parameter(torch.tensor(1.366e-4, dtype=torch.float32))
-        self.m = nn.Parameter(torch.tensor(0.068, dtype=torch.float32))
-        self.g = nn.Parameter(torch.tensor(9.81, dtype=torch.float32))
+        self.Jxx = nn.Parameter(torch.tensor(6.86e-5))
+        self.Jyy = nn.Parameter(torch.tensor(9.2e-5))  
+        self.Jzz = nn.Parameter(torch.tensor(1.366e-4))
+        self.m = nn.Parameter(torch.tensor(0.068))
+        self.g = nn.Parameter(torch.tensor(9.81))
         
         # True parameter values
         self.true_m = 0.068
